Move flight to components module
