+++ apps/fly-my-rockets/src/app/rockets/components/components.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SharedModule } from '../../shared/shared.module';
+import { FlightComponent } from './flight/flight.component';
+
+@NgModule({
+  declarations: [FlightComponent],
+  imports: [CommonModule, SharedModule],
+  exports: [FlightComponent]
+})
+export class ComponentsModule {}
