+++ apps/fly-my-rockets/src/app/rockets/components/flight/flight.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input } from '@angular/core';
-import { Flight } from '../rocket.model';
-import { FlightDialogComponent } from '../dialogs/flight-dialog/flight-dialog.component';
+import { Flight } from '../../rocket.model';
+import { FlightDialogComponent } from '../../dialogs/flight-dialog/flight-dialog.component';
 import { MatDialog } from '@angular/material/dialog';
-import { RocketService } from '../services/rocket/rocket.service';
-import { AreYouSureComponent } from '../../shared/are-you-sure/are-you-sure.component';
+import { RocketService } from '../../services/rocket/rocket.service';
+import { AreYouSureComponent } from '../../../shared/are-you-sure/are-you-sure.component';
 
 @Component({
   selector: 'fmr-flight',
